Correction of the method SubCircuit.check_nodes()
Change of the definition of Node to ensure the node name is always stored in lower case.
Converted the _external_nodes in SubCircuit into a tuple of Nodes.
Modified SubCircuit.check_nodes to use the Node name property, and a dictionary to verify if a node has more than one element connected.