Support for .nodeset type initial condition.
This is done by adding a _force parameter to initial_condition() method
of class CircuitSimulation. If _force is specified as False a .nodeset
statement will be generated. Advantage is that this value is also taken
into account for DC simulations. Difference is that this value is a hint
to the simulator at the start of DC convergence.