Don't delete DLL for ngspice version <=31
"this would be a good place to add one more fix for when user specifies
ngspice version <=31. In those versions, the DLL naming did not contain
a version number. …  if downloaded DLL's name is already 'ngspice.dll',
the unlink command causes the DLL to be deleted, and results in error.
… the code below adds a check to prevent deletion if the dll-vs
subdirectory only contains 'ngspice.dll'."

https://github.com/FabriceSalvaire/PySpice/pull/270#pullrequestreview-627574464

Co-authored-by: @mister-scott